requirements.txt
SQLAlchemy==1.4.22
psycopg2-binary==2.9.1
praw
transformers
